Started working on string class and added some code for loading shaders

+++ build.py
@@ -4,59 +4,74 @@ import shutil
 import tempfile
 from subprocess import call
 
-subDirs = ['utils', 'engine']
+subDirs = ['utils', 'engine', 'thirdparty']
 
 sourceFileLineStart = 'set(SOURCE_FILES'
 
+
+debug = False 
+
+def writeToFile(file, line):
+        if debug == True:
+                print(line[:-1])
+        else:
+                file.write(line)
+
 def build():
-	os.chdir('build')
+        os.chdir('build')
 
-	call('make -j30')
+        if debug == False:
+            call('make -j30')
 
-	os.chdir('..')
+        os.chdir('..')
 
 def writeToCmakeFile(sourceFiles):
-	cmakeSet = sourceFileLineStart
-	
-	for file in sourceFiles:
-		cmakeSet += ' ' + file
-	
-	cmakeSet += ')\n'
+        cmakeSet = sourceFileLineStart
 
-	filePath = 'CMakeLists.txt'
-	fh, absPath = tempfile.mkstemp()
-	with open(absPath, 'w') as newFile:
-		with open(filePath) as oldFile:
-			for line in oldFile:
-				if sourceFileLineStart in line:
-					newFile.write(cmakeSet)
-				else:
-					newFile.write(line)
+        for file in sourceFiles:
+                cmakeSet += ' ' + file
+        
+        cmakeSet += ')\n'
 
-	os.close(fh)
-	os.remove(filePath)
-	shutil.move(absPath, filePath)
+        filePath = 'CMakeLists.txt'
+        fh, absPath = tempfile.mkstemp()
+        with open(absPath, 'w') as newFile:
+                with open(filePath, 'r') as oldFile:
+                        for line in oldFile:
+                                if sourceFileLineStart in line:
+                                        writeToFile(newFile, cmakeSet)
+                                else:
+                                        writeToFile(newFile, line)
+
+        os.close(fh)
+        
+        if debug == False:
+                os.remove(filePath)
+                shutil.move(absPath, filePath)
 
 def parseSubdir(subDir):
-	if subDir == '':
-		return
+        if subDir == '':
+                return
 
-	os.chdir(subDir)
-	sourceFiles = glob.glob('*.cpp')
-	
-	writeToCmakeFile(sourceFiles)
+        if debug == True:
+                print('================================ Started writing ' + subDir + ' ===============================')
 
-	os.chdir('..')
+        os.chdir(subDir)
+        sourceFiles = glob.glob('*.cpp')
+        
+        writeToCmakeFile(sourceFiles)
+
+        os.chdir('..')
 
 def main():
-	sourceFiles = glob.glob('*.cpp')
-	writeToCmakeFile(sourceFiles)
+        sourceFiles = glob.glob('*.cpp')
+        writeToCmakeFile(sourceFiles)
 
-	for subDir in subDirs:
-		parseSubdir(subDir)
+        for subDir in subDirs:
+                parseSubdir(subDir)
 
-	build()
+        build()
 
 
 if __name__ == '__main__':
-	main()
+        main()
